+++ CHANGES.rst
@@ -13,6 +13,7 @@ Bug Fixes
   policies across both services. The endpoint now supports query parameter credentials to facilitate call directly
   within a web browser (not assuming JSON contents) and properly parses cookies from both policies to report more
   detailed error messages in case of failure.
+* Fix ``error`` and ``type`` representation when handling raised exception by ``evaluate_call`` utility function.
 
 `3.20.0 <https://github.com/Ouranosinc/Magpie/tree/3.20.0>`_ (2022-01-06)
 ------------------------------------------------------------------------------------